cosmossdk_average_network_blocktime/requirements.txt
requests
cosmospy
python-dateutil